Rationalised tab/window title handling...
In response to
<https://sourceforge.net/projects/roxterm/forums/forum/422638/topic/3935925>
Now the profile sets the tab name/title instead of the window title.
All tab titles/names set by the user may include %s to be substituted by
the terminal's title.

The current tab's full title is substituted into the window title which
has a default template of "ROXTerm: %s" which can be overridden from the
command line (but not from a profile).